Factor out pack to its own header
