+++ include/functional/utility.hpp
@@ -8,10 +8,10 @@
 
 #include "functional/detail/fwd_macro.hpp"
 #include "functional/detail/meta.hpp"
-#include "functional/detail/pack.hpp"
 #include "functional/detail/traits.hpp"
 
 #include <concepts>
+#include <functional>
 #include <type_traits>
 #include <utility>
 #include <variant>
@@ -26,101 +26,7 @@ template <typename T> [[nodiscard]] constexpr auto apply_const_lvalue(auto &&v) 
   return static_cast<apply_const_lvalue_t<T, decltype(v)>>(v);
 }
 
-template <typename... Ts> struct pack : detail::pack_base<std::index_sequence_for<Ts...>, Ts...> {
-  using _base = detail::pack_base<std::index_sequence_for<Ts...>, Ts...>;
-
-  template <typename Arg> using append_type = pack<Ts..., Arg>;
-
-  template <typename T>
-  [[nodiscard]] constexpr auto append(std::in_place_type_t<T>, auto &&...args) & noexcept -> append_type<T>
-    requires std::is_constructible_v<T, decltype(args)...>
-             && requires { static_cast<append_type<T>>(_base::template _append<T>(*this, FWD(args)...)); }
-  {
-    return {_base::template _append<T>(*this, FWD(args)...)};
-  }
-
-  template <typename T>
-  [[nodiscard]] constexpr auto append(std::in_place_type_t<T>, auto &&...args) const & noexcept -> append_type<T>
-    requires std::is_constructible_v<T, decltype(args)...>
-             && requires { static_cast<append_type<T>>(_base::template _append<T>(*this, FWD(args)...)); }
-  {
-    return {_base::template _append<T>(*this, FWD(args)...)};
-  }
-
-  template <typename T>
-  [[nodiscard]] constexpr auto append(std::in_place_type_t<T>, auto &&...args) && noexcept -> append_type<T>
-    requires std::is_constructible_v<T, decltype(args)...>
-             && requires { static_cast<append_type<T>>(_base::template _append<T>(std::move(*this), FWD(args)...)); }
-  {
-    return {_base::template _append<T>(std::move(*this), FWD(args)...)};
-  }
-
-  template <typename T>
-  [[nodiscard]] constexpr auto append(std::in_place_type_t<T>, auto &&...args) const && noexcept -> append_type<T>
-    requires std::is_constructible_v<T, decltype(args)...>
-             && requires { static_cast<append_type<T>>(_base::template _append<T>(std::move(*this), FWD(args)...)); }
-  {
-    return {_base::template _append<T>(std::move(*this), FWD(args)...)};
-  }
-
-  template <typename Arg>
-  [[nodiscard]] constexpr auto append(Arg &&arg) & noexcept -> append_type<Arg>
-    requires requires { static_cast<append_type<Arg>>(_base::template _append<Arg>(*this, FWD(arg))); }
-  {
-    return {_base::template _append<Arg>(*this, FWD(arg))};
-  }
-
-  template <typename Arg>
-  [[nodiscard]] constexpr auto append(Arg &&arg) const & noexcept -> append_type<Arg>
-    requires requires { static_cast<append_type<Arg>>(_base::template _append<Arg>(*this, FWD(arg))); }
-  {
-    return {_base::template _append<Arg>(*this, FWD(arg))};
-  }
-
-  template <typename Arg>
-  [[nodiscard]] constexpr auto append(Arg &&arg) && noexcept -> append_type<Arg>
-    requires requires { static_cast<append_type<Arg>>(_base::template _append<Arg>(std::move(*this), FWD(arg))); }
-  {
-    return {_base::template _append<Arg>(std::move(*this), FWD(arg))};
-  }
-
-  template <typename Arg>
-  [[nodiscard]] constexpr auto append(Arg &&arg) const && noexcept -> append_type<Arg>
-    requires requires { static_cast<append_type<Arg>>(_base::template _append<Arg>(std::move(*this), FWD(arg))); }
-  {
-    return {_base::template _append<Arg>(std::move(*this), FWD(arg))};
-  }
-
-  template <typename Fn>
-  [[nodiscard]] constexpr auto invoke(Fn &&fn, auto &&...args) & noexcept -> decltype(auto)
-    requires requires { _base::_invoke(*this, FWD(fn), FWD(args)...); }
-  {
-    return _base::_invoke(*this, FWD(fn), FWD(args)...);
-  }
-
-  template <typename Fn>
-  [[nodiscard]] constexpr auto invoke(Fn &&fn, auto &&...args) const & noexcept -> decltype(auto)
-    requires requires { _base::_invoke(*this, FWD(fn), FWD(args)...); }
-  {
-    return _base::_invoke(*this, FWD(fn), FWD(args)...);
-  }
-
-  template <typename Fn>
-  [[nodiscard]] constexpr auto invoke(Fn &&fn, auto &&...args) && noexcept -> decltype(auto)
-    requires requires { _base::_invoke(std::move(*this), FWD(fn), FWD(args)...); }
-  {
-    return _base::_invoke(std::move(*this), FWD(fn), FWD(args)...);
-  }
-
-  template <typename Fn>
-  [[nodiscard]] constexpr auto invoke(Fn &&fn, auto &&...args) const && noexcept -> decltype(auto)
-    requires requires { _base::_invoke(std::move(*this), FWD(fn), FWD(args)...); }
-  {
-    return _base::_invoke(std::move(*this), FWD(fn), FWD(args)...);
-  }
-};
-
-template <typename... Args> pack(Args &&...args) -> pack<Args...>;
+template <typename... Ts> struct pack;
 
 namespace detail {
 template <typename... Ts> constexpr bool _is_some_pack = false;
@@ -131,10 +37,22 @@ template <typename... Ts> constexpr bool _is_some_pack<::fn::pack<Ts...> const &
 template <typename T>
 concept some_pack = detail::_is_some_pack<T &>;
 
+template <typename... Ts> struct sum;
+
+namespace detail {
+template <typename... Ts> constexpr bool _is_sum = false;
+template <typename... Ts> constexpr bool _is_sum<sum<Ts...> &> = true;
+template <typename... Ts> constexpr bool _is_sum<sum<Ts...> const &> = true;
+} // namespace detail
+
+template <typename T>
+concept some_sum = detail::_is_sum<T &>;
+
 template <typename... Ts> struct overload final : Ts... {
   using Ts::operator()...;
 };
 template <typename... Ts> overload(Ts const &...) -> overload<Ts...>;
+
 } // namespace fn
 
 #endif // INCLUDE_FUNCTIONAL_UTILITY
